requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0

# Database (Async SQLite example)
aiosqlite>=0.19.0

# Vector DB (Chroma example)
chromadb>=0.4.24

# Embeddings
sentence-transformers>=2.6.0

# PDF Processing (Example - PyMuPDF)
pymupdf>=1.24.0

# Optional: For SQLAlchemy ORM (if chosen later)
# sqlalchemy[asyncio]>=2.0.0
# alembic>=1.13.0 # For migrations

# Optional: For Celery (if chosen later)
# celery>=5.3.0
# redis>=5.0.0 # Example broker
